requirements.txt
streamlit
pandas
openai
python-dotenv
plotly
langchain-core
langgraph
groq
openpyxl